Fixed some test issues
