+++ src/Crash.Common/Communications/CrashServer.cs
@@ -1,6 +1,4 @@
 ﻿using System.Diagnostics;
-using System.IO;
-using System.Threading;
 
 using Crash.Common.Document;
 using Crash.Common.Events;
